[fix]관리자 및 유저 RESTful API 설정
