+++ eoullim-Project/src/main/java/org/example/eoullimback/user_auth/auth/AuthController.java
@@ -12,7 +12,6 @@ import org.example.eoullimback.user_auth.user.MailService;
 import org.example.eoullimback.user_auth.user.User;
 import org.example.eoullimback.user_auth.user.UserService;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
