+++ eoullim-Project/src/main/java/org/example/eoullimback/user_auth/user/UserController.java
@@ -19,7 +19,7 @@ public class UserController {
 
     private final UserService userService;
 
-    @GetMapping("/oauth/kakao")
+    @GetMapping("/kakao")
     public String kakaoCallback(@RequestParam(name = "code") String code, HttpSession session) {
 
         try {
